requirements.txt
asgiref==3.5.2
certifi==2022.5.18
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==0.5.0
dj-rest-auth==2.2.4
Django==4.0.5
django-allauth==0.50.0
django-countries==7.3.2
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.20.0
gcloud==0.18.3
googleapis-common-protos==1.56.1
gunicorn==20.1.0
httplib2==0.20.4
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.3
Markdown==3.3.7
MarkupSafe==2.1.1
oauth2client==4.1.3
oauthlib==3.2.0
packaging==21.3
Pillow==9.1.1
protobuf==4.21.0rc1
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.14.1
PyJWT==2.4.0
pyotp==2.6.0
pyparsing==3.0.9
Pyrebase4==4.5.0
python-dotenv==0.20.0
python-jwt==3.3.2
python3-openid==3.2.0
pytz==2022.1
PyYAML==5.3.1
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.2.0
tzdata==2022.1
uritemplate==3.0.1
urllib3==1.26.9
whitenoise==6.1.0
wrapt==1.14.1
